chore: Update 'webdriver/index' file to ts
